requirements.txt
django~=5.0.2
python-dotenv~=1.0.1

# psycopg2-binary
# mysqlclient

pandas~=2.2.0
djangorestframework~=3.14.0